Use duckdb by default

+++ models/int_order_payments_pivoted.sql
@@ -18,10 +18,10 @@ order_payments as (
         order_id,
 
         {% for payment_method in payment_methods -%}
-        sum(case when payment_method = '{{ payment_method }}' then amount else 0 end) as {{ payment_method }}_amount,
+        sum(case when payment_method = '{{ payment_method }}' then amount else 0 end)::bigint as {{ payment_method }}_amount,
         {% endfor -%}
 
-        sum(amount) as total_amount
+        sum(amount)::bigint as total_amount
 
     from payments
